requirements.txt
ipython==8.12.3
matplotlib==3.5.3
networkx==2.6.3
numpy==1.22.4
pandas==1.3.5
Pillow==9.4.0
scikit_learn==1.1.2
scipy==1.10.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
